Revert "remove path simplification (causes problems in multi-threaded)"
This reverts commit 25517cc7f80fe5a064d6ef5d4c5b646516c49cc6.

* src/path.lisp: (*use-previous-path-instruction-hack*): set to nil to
get the upstream behaviour i.e. without any path simplification.

The simplification behaviour is desired to produce path commands which
involve repetition. Consider the following path which produces a V
shape.

```
<svg viewBox="0 0 700 700" xmlns="http://www.w3.org/2000/svg">
 <path d="M20 20m45.00 0 117.50 325 25 0 117.50 -325 45.00 0 -148.08 350 -53.85 0 -148.08 -350z"/>
</svg>
```

The entire path is produced by the close-path `Z' command which is
defined to operate on the previous command which is a `m' command with

Restoring path simplification allows the cl-svg user to proudce that
path with a lisp fragment like

```
 (svg:path
  (svg:move-to      x y)
  (svg:move-to-r    ...)
  (svg:move-to-r    ...)
  (svg:close-path))
```

Without path simplification the lisp code fragment would produce

```
 <path d="M20 20m45.00 0m117.50 325m25 0m117.50 -325m45.00 0m-148.08 350m-53.85 0m-148.08 -350z"/>
```

which renders /nothing/. The user is required to write

```
 (svg:path
  (svg:move-to      x y)
  (svg:line-to-r    ...)
  (svg:line-to-r    ...)
  (svg:close-path))
```

to produce

```
<svg viewBox="0 0 700 700" xmlns="http://www.w3.org/2000/svg">
 <path d="M20 20l45.00 0l117.50 325l25 0l117.50 -325l45.00 0l-148.08 350l-53.85 0l-148.08 -350z"/>
</svg>
```
and the path would still not be closed.

Perhaps move-to-r and line-to-r commands can take additional unlimited
pairs of (x y) arguments but that does not make for expressing cleaner
lisp code (which computes those points), and automatic path
simplification using the above idiom seems to be the better solution